renaming directory => site directory
