+++ COMETwebapp/Extensions/ServiceCollectionExtensions.cs
@@ -30,7 +30,7 @@ namespace COMETwebapp.Extensions
     using COMETwebapp.Services.SubscriptionService;
     using COMETwebapp.Utilities;
     using COMETwebapp.ViewModels.Components.BookEditor;
-    using COMETwebapp.ViewModels.Components.Directory.DomainsOfExpertise;
+    using COMETwebapp.ViewModels.Components.SiteDirectory.DomainsOfExpertise;
     using COMETwebapp.ViewModels.Components.ModelDashboard;
     using COMETwebapp.ViewModels.Components.ModelDashboard.ParameterValues;
     using COMETwebapp.ViewModels.Components.ModelEditor;
