+++ COMETwebapp/Pages/SiteDirectory/DirectoryPage.razor
@@ -14,7 +14,7 @@ Copyright (c) 2023-2024 RHEA System S.A.
     You should have received a copy of the GNU Affero General Public License
     along with this program. If not, see http://www.gnu.org/licenses/.
 ------------------------------------------------------------------------------->
-@attribute [Route(WebAppConstantValues.DirectoryPage)]
+@attribute [Route(WebAppConstantValues.SiteDirectoryPage)]
 @attribute [Authorize]
 @using COMETwebapp.Utilities
