+++ COMETwebapp/ViewModels/Components/SiteDirectory/DomainsOfExpertise/IDomainsOfExpertiseTableViewModel.cs
@@ -22,12 +22,12 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace COMETwebapp.ViewModels.Components.Directory.DomainsOfExpertise
+namespace COMETwebapp.ViewModels.Components.SiteDirectory.DomainsOfExpertise
 {
     using CDP4Common.SiteDirectoryData;
 
     using COMETwebapp.ViewModels.Components.Common.DeprecatableDataItem;
-    using COMETwebapp.ViewModels.Components.Directory.Rows;
+    using COMETwebapp.ViewModels.Components.SiteDirectory.Rows;
 
     /// <summary>
     /// View model used to manage <see cref="DomainOfExpertise" />
